test(token-ops): merge operations and pending e2e specs into actions
Fold the separate `operations.spec.ts` and `pending.spec.ts` into one `actions.spec.ts` with two describes (propose form + pending queue), matching the merged Actions tab and the one-file-per-tab convention the other specs already follow.

Refresh the docblock to describe a single tab with two stacked panels and update the CLAUDE.md e2e reference.
