+++ example-apps/token-ops/CLAUDE.md
@@ -110,7 +110,7 @@ Vitest ([test/](test/), `test/**/*.test.{ts,tsx}`, default env `node`, jsdom via
 - WIF sign-in: the mnemonic-vs-WIF heuristic ([detectSecretShape.test.ts](test/detectSecretShape.test.ts)), the key-validation branches of `resolveIdentityFromWif`/`loginWithPrivateKey` with `@dashevo/evo-sdk` mocked ([loginWithPrivateKey.test.ts](test/loginWithPrivateKey.test.ts)), and the auth-tuple adapter ([keyManagerFromKey.test.ts](test/keyManagerFromKey.test.ts))
 - component tests for `LoginModal` / `TopNav` / `OverviewView` / `GovernanceView` / `ActionsView` / `ProposeActionPanel` / `PendingActionsView` (mock the dash modules + `useSession`; `ActionsView.test.tsx` asserts one shared governance fetch and refresh propagation; `OverviewView.test.tsx` covers authority aggregation, personalized counts/membership, navigation, loading/errors, and lazy balance inspection; `LoginModal.test.tsx` also covers the mnemonic-only "advanced settings" toggle)
 
-E2E ([test/e2e/](test/e2e/), Playwright, port 5184, real testnet, chromium only, serial): the suite is **read-only** and performs no chain writes. `browse.spec.ts` is the boot/nav/mobile smoke baseline (nav expects **Dashboard** and **Settings**; the Dashboard assertion checks its governance summary and token context); [operations.spec.ts](test/e2e/operations.spec.ts), [settings.spec.ts](test/e2e/settings.spec.ts), [governance.spec.ts](test/e2e/governance.spec.ts), and [overview.spec.ts](test/e2e/overview.spec.ts) broaden per-view read coverage (read-only notices/gating, the live contract-ID resolver, Groups search/filter/sort/expand, the Access matrices, the lazy identity-balance inspector). These are **ungated** and always run — no credentials needed — mirroring the fully-read-only dashrate suite. The one gated spec, [authenticated-readonly.spec.ts](test/e2e/authenticated-readonly.spec.ts), signs in (identity index 0) only to observe the authenticated read state and sign out, and asserts the **mnemonic and WIF sign-in paths resolve to the same identity**; it uses `loginAs` (mnemonic + "Show advanced settings" identity index) and `loginWithWif` (the WIF is derived from the same mnemonic via `deriveWifFromMnemonic` — DIP-13 identity index 0, CRITICAL auth key 2 — so one funded mnemonic drives both paths), both in [fixtures.ts](test/e2e/fixtures.ts), and `test.skip(!HAS_MNEMONIC, …)` cleanly when `PLATFORM_MNEMONIC` (repo-root `.env`) is unset. Data-dependent reads (e.g. the inspector's live balance lookup, keyed off the three `VITE_TOKEN_OPS_MEMBER_*_ID` vars / `HAS_GROUP_IDENTITIES`) `test.skip(true, …)` when the live state has nothing to assert on. **Write flows (propose / co-sign→execute round-trip / transfer / reassign / append group / register contract) are deferred to a future pass** — see the header comment in [operations.spec.ts](test/e2e/operations.spec.ts). Any chain-mutating e2e must avoid the irreversible `destroyFrozen` path unless behind an explicit manual flag.
+E2E ([test/e2e/](test/e2e/), Playwright, port 5184, real testnet, chromium only, serial): the suite is **read-only** and performs no chain writes. `browse.spec.ts` is the boot/nav/mobile smoke baseline (nav expects **Dashboard** and **Settings**; the Dashboard assertion checks its governance summary and token context); [actions.spec.ts](test/e2e/actions.spec.ts) (two describes — propose form + pending queue, the two panels of the merged Actions tab), [settings.spec.ts](test/e2e/settings.spec.ts), [governance.spec.ts](test/e2e/governance.spec.ts), and [overview.spec.ts](test/e2e/overview.spec.ts) broaden per-view read coverage (read-only notices/gating, the live contract-ID resolver, Groups search/filter/sort/expand, the Access matrices, the lazy identity-balance inspector). These are **ungated** and always run — no credentials needed — mirroring the fully-read-only dashrate suite. The one gated spec, [authenticated-readonly.spec.ts](test/e2e/authenticated-readonly.spec.ts), signs in (identity index 0) only to observe the authenticated read state and sign out, and asserts the **mnemonic and WIF sign-in paths resolve to the same identity**; it uses `loginAs` (mnemonic + "Show advanced settings" identity index) and `loginWithWif` (the WIF is derived from the same mnemonic via `deriveWifFromMnemonic` — DIP-13 identity index 0, CRITICAL auth key 2 — so one funded mnemonic drives both paths), both in [fixtures.ts](test/e2e/fixtures.ts), and `test.skip(!HAS_MNEMONIC, …)` cleanly when `PLATFORM_MNEMONIC` (repo-root `.env`) is unset. Data-dependent reads (e.g. the inspector's live balance lookup, keyed off the three `VITE_TOKEN_OPS_MEMBER_*_ID` vars / `HAS_GROUP_IDENTITIES`) `test.skip(true, …)` when the live state has nothing to assert on. **Write flows (propose / co-sign→execute round-trip / transfer / reassign / append group / register contract) are deferred to a future pass** — see the header comment in [actions.spec.ts](test/e2e/actions.spec.ts). Any chain-mutating e2e must avoid the irreversible `destroyFrozen` path unless behind an explicit manual flag.
 
 ## Gotchas
